+++ library/src/main/java/com/twilio/video/VideoTrack.java
@@ -94,9 +94,10 @@ public abstract class VideoTrack implements Track {
     }
 
     /**
-     * The list of renderers receiving video from this video track.
+     * The list of renderers receiving video from this video track. An empty list will be returned
+     * if the video track has been released.
      */
-    public List<VideoRenderer> getRenderers() {
+    public synchronized List<VideoRenderer> getRenderers() {
         return new ArrayList<>(videoRenderersMap.keySet());
     }