+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Improvements
 
 - Moved pre-defined aspect ratios from `VideoConstraints` class to `AspectRatio` class.
 - Local audio, video, and data tracks return their track IDs for `getName` if no name was specified.
+- Improved threading contract.
     
 Bug Fixes
 
@@ -235,6 +236,19 @@ Known issues
 
 - Network handoff, and subsequent connection renegotiation is not supported for IPv6 networks [#72](https://github.com/twilio/video-quickstart-android/issues/72)
 - Participant disconnect event can take up to 120 seconds to occur [#80](https://github.com/twilio/video-quickstart-android/issues/80) [#73](https://github.com/twilio/video-quickstart-android/issues/73)
+
+####1.3.5
+
+Improvements
+
+- Improved threading contract. 
+
+Known issues
+
+- Network handoff, and subsequent connection renegotiation is not supported for IPv6 networks [#72](https://github.com/twilio/video-quickstart-android/issues/72)
+- VP8 is the only supported codec [#71](https://github.com/twilio/video-quickstart-android/issues/71)
+- Participant disconnect event can take up to 120 seconds to occur [#80](https://github.com/twilio/video-quickstart-android/issues/80) [#73](https://github.com/twilio/video-quickstart-android/issues/73)
+- LocalParticipant release method is public [#132](https://github.com/twilio/video-quickstart-android/issues/132)
 
 ####1.3.4